Delete confirmation alert implemented

+++ src/app/pages/view-all-emp/view-all-emp.component.ts
@@ -1,6 +1,7 @@
 import { NgFor } from '@angular/common';
 import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-view-all-emp',
@@ -27,8 +28,41 @@ export class ViewAllEmpComponent implements OnInit {
   }
 
   public deleteEmployee(employee:any){
-    this.http.delete(`http://localhost:8080/emp-controller/delete-emp/${employee.id}`, {responseType:'text'}).subscribe(res => {
-      this.getAllEmp();
-    })    
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: "btn btn-success",
+        cancelButton: "btn btn-danger"
+      },
+      buttonsStyling: false
+    });
+    swalWithBootstrapButtons.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "No, cancel!",
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.http.delete(`http://localhost:8080/emp-controller/delete-emp/${employee.id}`, {responseType:'text'}).subscribe(res => {
+          this.getAllEmp();
+        })  
+        swalWithBootstrapButtons.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success"
+        });
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire({
+          title: "Cancelled",
+          text: "Your imaginary file is safe :)",
+          icon: "error"
+        });
+      }
+    });      
   }
 }
